finished Java version of Prim's MST Problem

+++ 3-Greedy Algorithms, Minimum Spanning Tree, and Dynamic Programming/1-Greedy Algorithm/3-Minimum Spanning Tree Problem/Java/undirected_graph/Vertex.java
@@ -16,7 +16,7 @@ class Vertex extends AbstractVertex implements Comparable<Vertex> {
     /**
      * Default minimum cost of the incident edge from the spanned vertices (X).
      */
-    private static final double DEFAULT_MIN_INCIDENT_COST = Integer.MAX_VALUE;
+    private static final int DEFAULT_MIN_INCIDENT_COST = Integer.MAX_VALUE;
 
     /**
      * Frequency of neighbors.
